HEEDLS-532 Move view component models into folder
