+++ DigitalLearningSolutions.Web/ViewComponents/ErrorSummaryViewComponent.cs
@@ -2,7 +2,7 @@
 {
     using System.Collections.Generic;
     using System.Linq;
-    using DigitalLearningSolutions.Web.ViewModels.Common;
+    using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents;
     using Microsoft.AspNetCore.Mvc;
 
     public class ErrorSummaryViewComponent : ViewComponent
