+++ DigitalLearningSolutions.Web/ViewComponents/TermsConditionsViewComponent.cs
@@ -1,6 +1,6 @@
 ﻿namespace DigitalLearningSolutions.Web.ViewComponents
 {
-    using DigitalLearningSolutions.Web.ViewModels.Common;
+    using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents;
     using Microsoft.AspNetCore.Mvc;
 
     public class TermsConditionsViewComponent : ViewComponent
