+++ DigitalLearningSolutions.Web/ViewModels/Common/ViewComponents/TermsConditionsViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.Common
+﻿namespace DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 {
     public class TermsConditionsViewModel
     {
