+++ DigitalLearningSolutions.Web/ViewModels/Common/ViewComponents/TextAreaViewModel.cs
@@ -1,21 +1,22 @@
-﻿namespace DigitalLearningSolutions.Web.ViewModels.Common
+﻿namespace DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 {
     using System.Collections.Generic;
     using System.Linq;
 
-    public class TextInputViewModel
+    public class TextAreaViewModel
     {
-        public TextInputViewModel(
+        public TextAreaViewModel
+        (
             string id,
             string name,
             string label,
             string? value,
-            string type,
+            int rows,
             bool spellCheck,
-            string? autocomplete,
             IEnumerable<string> errorMessages,
             string? cssClass = null,
-            string? hintText = null
+            string? hintText = null,
+            int? characterCount = null
         )
         {
             var errorMessageList = errorMessages.ToList();
@@ -25,10 +26,10 @@
             Name = name;
             Label = label;
             Value = value;
-            Type = type;
+            Rows = rows;
             SpellCheck = spellCheck;
-            Autocomplete = autocomplete;
             HintText = hintText;
+            CharacterCount = characterCount;
             ErrorMessages = errorMessageList;
             HasError = errorMessageList.Any();
         }
@@ -38,11 +39,11 @@
         public string Name { get; set; }
         public string Label { get; set; }
         public string? Value { get; set; }
-        public string Type { get; set; }
+        public int Rows { get; set; }
         public bool SpellCheck { get; set; }
-        public string? Autocomplete { get; set; }
         public string? HintText { get; set; }
+        public int? CharacterCount { get; set; }
         public IEnumerable<string> ErrorMessages { get; set; }
-        public readonly bool HasError;
+        public bool HasError { get; set; }
     }
 }
