+++ DigitalLearningSolutions.Web/Views/Shared/Components/ActionLink/Default.cshtml
@@ -1,4 +1,4 @@
-@using DigitalLearningSolutions.Web.ViewModels.Common
+@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model LinkViewModel
 
 <div class="nhsuk-action-link">
