+++ DigitalLearningSolutions.Web/Views/Shared/Components/CentreContactInfo/Default.cshtml
@@ -1,4 +1,4 @@
-@using DigitalLearningSolutions.Web.ViewModels.Common
+@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model CentreContactInfoViewModel
 
 <link rel="stylesheet" href="@Url.Content("~/css/shared/centreInfo.css")" asp-append-version="true">
