+++ DigitalLearningSolutions.Web/Views/Shared/Components/ErrorSummary/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model ErrorSummaryViewModel
 
 <div class="nhsuk-error-summary" aria-labelledby="error-summary-title" role="alert" tabindex="-1">
