+++ DigitalLearningSolutions.Web/Views/Shared/Components/Logo/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model LogoViewModel
 @if (Model.Logo?.LogoUrl != null)
 {
