+++ DigitalLearningSolutions.Web/Views/Shared/Components/NumberOfAdministrators/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model NumberOfAdministratorsViewModel
 
 <link rel="stylesheet" href="@Url.Content("~/css/trackingSystem/numberOfAdministratorsViewComponent.css")" asp-append-version="true">
