+++ DigitalLearningSolutions.Web/Views/Shared/Components/SearchBox/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model SearchBoxViewModel
 
 <div class="@Model.Class" id="search">
