+++ DigitalLearningSolutions.Web/Views/Shared/Components/SelectList/Default.cshtml
@@ -1,4 +1,4 @@
-@using DigitalLearningSolutions.Web.ViewModels.Common
+@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model SelectListViewModel
 
 <div class="nhsuk-form-group @(Model.HasError ? "nhsuk-form-group--error" : "")">
