+++ DigitalLearningSolutions.Web/Views/Shared/Components/SideMenuLink/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model SideMenuLinkViewModel
 
 @if (Model.IsCurrentPage) {
