+++ DigitalLearningSolutions.Web/Views/Shared/Components/SideMenuLinkWithHref/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model SideMenuLinkWithHrefViewModel
 
 @if (Model.IsCurrentPage) {
