+++ DigitalLearningSolutions.Web/Views/Shared/Components/TermsConditions/Default.cshtml
@@ -1,4 +1,4 @@
-﻿@using DigitalLearningSolutions.Web.ViewModels.Common
+﻿@using DigitalLearningSolutions.Web.ViewModels.Common.ViewComponents
 @model TermsConditionsViewModel
 
 <div class="nhsuk-form-group @(Model.HasError ? "nhsuk-form-group--error" : "")">
